test: Address code review findings on the chaos suite
Centralize the duplicated per-file helpers (proxy/port mapping,
connected-master lookup, retried connection, healthy-replica wait)
into the InteractsWithToxiproxy trait so single-file runs keep working
without diverging copies. Fail fast when toxiproxy rejects a disable,
enable, or toxic removal instead of letting tests hang on a silent
no-op. Add a deterministic READONLY test that primes a stale node
address cache against the demoted master, deriving the Sentinel
service name from config instead of hardcoding it, and tighten the
README wording accordingly.
